requirements.txt
pandas
sqlalchemy
psycopg2-binary
kafka-python
scikit-learn
joblib
streamlit
plotly
streamlit-lottie
requests
aiosmtpd